+++ app/src/main/kotlin/io/hammerhead/karoocriticalpower/datatypes/CriticalPower5sDataType.kt
@@ -1,6 +1,7 @@
 package io.hammerhead.karoocriticalpower.datatypes
 
 import io.hammerhead.karooext.KarooSystemService
+import io.hammerhead.karoocriticalpower.PowerBufferManager
 import io.hammerhead.karoocriticalpower.data.PowerCurveRepository
 import io.hammerhead.karoocriticalpower.data.PrTimeframe
 
@@ -10,6 +11,7 @@ import io.hammerhead.karoocriticalpower.data.PrTimeframe
 class CriticalPower5sDataType(
     extensionId: String,
     karooSystem: KarooSystemService,
+    bufferManager: PowerBufferManager,
     powerCurveRepository: PowerCurveRepository,
     showPrComparison: () -> Boolean,
     getPrTimeframe: () -> PrTimeframe,
@@ -19,6 +21,7 @@ class CriticalPower5sDataType(
     karooSystem = karooSystem,
     durationSeconds = 5,
     typeIdSuffix = "5s",
+    bufferManager = bufferManager,
     powerCurveRepository = powerCurveRepository,
     showPrComparison = showPrComparison,
     getPrTimeframe = getPrTimeframe,